# samples/sample-openapi.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "Sample API",
		"version": "1.0.0",
		"description": "A sample API for testing oas2msw"
	},
	"paths": {
		"/users": {
			"get": {
				"summary": "Get users",
				"responses": {
					"200": {
						"description": "Successful response",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"items": {
											"type": "array",
											"items": {
												"$ref": "#/components/schemas/User"
											}
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request"
					},
					"500": {
						"description": "Internal server error"
					}
				}
			},
			"post": {
				"summary": "Create user",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUserRequest"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "User created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				}
			}
		},
		"/users/{id}": {
			"get": {
				"summary": "Get user by ID",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"description": "User ID"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					},
					"404": {
						"description": "User not found"
					}
				}
			}
		}
	},
	"components": {
		"schemas": {
			"User": {
				"type": "object",
				"properties": {
					"id": {
						"type": "integer",
						"description": "User ID"
					},
					"name": {
						"type": "string",
						"maxLength": 100
					},
					"email": {
						"type": "string",
						"format": "email"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"isActive": {
						"type": "boolean"
					}
				},
				"required": ["id", "name", "email"]
			},
			"CreateUserRequest": {
				"type": "object",
				"properties": {
					"name": {
						"type": "string",
						"maxLength": 100
					},
					"email": {
						"type": "string",
						"format": "email"
					}
				},
				"required": ["name", "email"]
			}
		}
	}
}